chore(linky-note): rename project
