+++ README.md
@@ -1,16 +1,16 @@
-# marko-backlinks
+# linky-note
 
 <div align="center">
 
-[![Build status](https://github.com/jb-delafosse/marko-backlinks/workflows/build/badge.svg?branch=master&event=push)](https://github.com/jb-delafosse/marko-backlinks/actions?query=workflow%3Abuild)
-[![Python Version](https://img.shields.io/pypi/pyversions/marko-backlinks.svg)](https://pypi.org/project/marko-backlinks/)
-[![Dependencies Status](https://img.shields.io/badge/dependencies-up%20to%20date-brightgreen.svg)](https://github.com/jb-delafosse/marko-backlinks/pulls?utf8=%E2%9C%93&q=is%3Apr%20author%3Aapp%2Fdependabot)
+[![Build status](https://github.com/jb-delafosse/linky-note/workflows/build/badge.svg?branch=master&event=push)](https://github.com/jb-delafosse/linky-note/actions?query=workflow%3Abuild)
+[![Python Version](https://img.shields.io/pypi/pyversions/linky-note.svg)](https://pypi.org/project/linky-note/)
+[![Dependencies Status](https://img.shields.io/badge/dependencies-up%20to%20date-brightgreen.svg)](https://github.com/jb-delafosse/linky-note/pulls?utf8=%E2%9C%93&q=is%3Apr%20author%3Aapp%2Fdependabot)
 
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Security: bandit](https://img.shields.io/badge/security-bandit-green.svg)](https://github.com/PyCQA/bandit)
-[![Pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/jb-delafosse/marko-backlinks/blob/master/.pre-commit-config.yaml)
-[![Semantic Versions](https://img.shields.io/badge/%F0%9F%9A%80-semantic%20versions-informational.svg)](https://github.com/jb-delafosse/marko-backlinks/releases)
-[![License](https://img.shields.io/github/license/jb-delafosse/marko-backlinks)](https://github.com/jb-delafosse/marko-backlinks/blob/master/LICENSE)
+[![Pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white)](https://github.com/jb-delafosse/linky-note/blob/master/.pre-commit-config.yaml)
+[![Semantic Versions](https://img.shields.io/badge/%F0%9F%9A%80-semantic%20versions-informational.svg)](https://github.com/jb-delafosse/linky-note/releases)
+[![License](https://img.shields.io/github/license/jb-delafosse/linky-note)](https://github.com/jb-delafosse/linky-note/blob/master/LICENSE)
 
 </div>
 
@@ -41,7 +41,7 @@ Using git and this converter as a pre-commit, I can easily do all of this ! 🚀
 - Understands both  Wikilinks and Markdown links
 - Can use a reference system based on Title as unique Keys or filename as unique key
 - Can convert wikilinks to markdown links and reciprocally   
-- All this, entirely configurable through a simple stepper using `marko-backlinks init` command
+- All this, entirely configurable through a simple stepper using `linky-note init` command
 
 ![init](img/init.png)
 
@@ -53,25 +53,25 @@ Using git and this converter as a pre-commit, I can easily do all of this ! 🚀
 Considering you already have python available. You can simply add th
 
 ```bash
-pip install --user marko-backlinks
+pip install --user linky_note
 ```
 
 Then you can see all the option of the CLI using
 
 ```bash
-marko-backlinks --help
+linky_note --help
 ```
 
 It is advised to start by configuring the CLI using
 
 ```bash
-marko-backlinks init
+linky_note init
 
 ```
 You can then apply the conversion 
 
 ```bash
-marko-backlinks apply <INPUT_DIR> --output-dir <OUTPUT_DIR> 
+linky_note apply <INPUT_DIR> --output-dir <OUTPUT_DIR> 
 
 ```
 
@@ -89,35 +89,35 @@ repository.
 
 ```yaml
 repos:
--   repo: https://github.com/jb-delafosse/marko-backlinks
-    rev: v0.2.3
+-   repo: https://github.com/jb-delafosse/linky_note
+    rev: v0.4.0
     hooks:
-      - id: marko-backlinks apply
+      - id: linky_note apply
         args: ['directory-containing-my-markdown']
 ```
 and install the hook using `pre-commit install`
 
 
-You should also run `marko-backlinks init` at the root of your repo to configure marko-backlinks
+You should also run `linky init` at the root of your repo to configure linky-note
 
 </details>
 
 ## 🛡 License
 
-[![License](https://img.shields.io/github/license/jb-delafosse/marko-backlinks)](https://github.com/jb-delafosse/marko-backlinks/blob/master/LICENSE)
+[![License](https://img.shields.io/github/license/jb-delafosse/linky-note)](https://github.com/jb-delafosse/linky-note/blob/master/LICENSE)
 
-This project is licensed under the terms of the `MIT` license. See [LICENSE](https://github.com/jb-delafosse/marko-backlinks/blob/master/LICENSE) for more details.
+This project is licensed under the terms of the `MIT` license. See [LICENSE](https://github.com/jb-delafosse/linky-note/blob/master/LICENSE) for more details.
 
 ## 📃 Citation
 
 ```
-@misc{marko-backlinks,
+@misc{linkynote,
   author = {jb-delafosse},
-  title = {Awesome `marko-backlinks` is a Python cli/package created with https://github.com/TezRomacH/python-package-template},
+  title = {Awesome `linkynote` is a Python cli/package created with https://github.com/TezRomacH/python-package-template},
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
-  howpublished = {\url{https://github.com/jb-delafosse/marko-backlinks}}
+  howpublished = {\url{https://github.com/jb-delafosse/linkynote}}
 }
 ```
