--- cookiecutter-config-file.yml
@@ -1,10 +1,0 @@
-# This file contains values from Cookiecutter
-
-default_context:
-  project_name: "marko-backlinks"
-  project_description: "Awesome `marko-backlinks` is a Python cli/package created with https://github.com/TezRomacH/python-package-template"
-  organization: "jb-delafosse"
-  license: "MIT"
-  version: "0.1.0"
-  github_name: "jb-delafosse"
-  email: "jean-baptiste@lumapps.com"
