+++ linky_note/adapters/markdown/marko_ext/marko_builder.py
@@ -1,4 +1,4 @@
-from typing import List, Union, Optional
+from typing import List, Optional, Union
 
 from marko.block import BlankLine, Heading
 from marko.block import List as MdList
