+++ linky_note/adapters/references_db/query_builders.py
@@ -1,8 +1,8 @@
 from typing import Optional
 
-from marko_backlinks.adapters.references_db import tables
-from marko_backlinks.dto.dto import Note, Reference, ReferenceBy
-from marko_backlinks.interfaces.references_db import (
+from linky_note.adapters.references_db import tables
+from linky_note.dto.dto import Note, Reference, ReferenceBy
+from linky_note.interfaces.references_db import (
     GetNoteByTitleQuery,
     GetNoteResponse,
     GetReferencesResponse,
